pkg/fabsdk/factory/defclient/orgfactory.go
/*
Copyright SecureKey Technologies Inc. All Rights Reserved.

SPDX-License-Identifier: Apache-2.0
*/

package defclient

import (
	"github.com/hyperledger/fabric-sdk-go/pkg/context/api/core"
	"github.com/hyperledger/fabric-sdk-go/pkg/context/api/fab"
	"github.com/hyperledger/fabric-sdk-go/pkg/fab/identitymgr"
)

// OrgClientFactory represents the default org provider factory.
type OrgClientFactory struct{}

// NewOrgClientFactory returns the default org provider factory.
func NewOrgClientFactory() *OrgClientFactory {
	f := OrgClientFactory{}
	return &f
}

// New returns a new default implementation of the credential manager
func (f *OrgClientFactory) New(orgName string, config core.Config, cryptoProvider core.CryptoSuite) (fab.IdentityManager, error) {
	return identitymgr.New(orgName, config, cryptoProvider)
}
